Test refactor #2
